Massive commit to get building on xcode 10
Lots of plugins needed upgrading as they had died in the intervening
two years - most importantly the SQLite one has changed name/namespace
but seems to be broadly speaking compatible. At this stage, I can
deploy to emulator and see a map but menu bar remains blank - unclear
if this is because login server on brightbox is down or because more
code fixes are required. README has been upadted.